scheduler/requirements.txt
web3
pyspark
python-dotenv
apache-airflow